Add recommendations as separate cards
